backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.3
django-cors-headers==4.9.0
channels==4.0.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
paystackapi==2.1.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
reportlab>=3.6.0
gunicorn==23.0.0
psutil==5.9.5
openpyxl==3.1.2
daphne>=4.0.0
channels>=4.0.0
openpyxl>=3.1.2